Move tests of flask_common.formfields to tests.test_formfields
