+++ tests/test_legacy.py
@@ -1,24 +1,19 @@
 # -*- coding: utf-8 -*-
 
-import datetime
 import random
 import string
 import unittest
 
-from dateutil.tz import tzutc
 from flask import Flask
 from mongoengine import connection
 from mongoengine.errors import DoesNotExist
 from mongoengine.fields import (ReferenceField, SafeReferenceField,
                                 SafeReferenceListField, StringField,
                                 IntField)
-from werkzeug.datastructures import MultiDict
-from wtforms import Form
 
 from flask_mongoengine import MongoEngine
 from flask_common.crypto import aes_generate_key
 from flask_common.declenum import DeclEnum
-from flask_common.formfields import BetterDateTimeField
 from flask_common.mongo import (
     EncryptedStringField,
     LowerEmailField,
@@ -53,31 +48,6 @@ class Book(db.Document):
 
 class Author(db.Document):
     books = SafeReferenceListField(ReferenceField(Book))
-
-
-class FormFieldTestCase(unittest.TestCase):
-
-    def test_datetime_field(self):
-        class TestForm(Form):
-            date = BetterDateTimeField()
-
-        form = TestForm(MultiDict({'date': ''}))
-        self.assertTrue(form.validate())
-        self.assertEqual(form.data['date'], None)
-
-        form = TestForm(MultiDict({'date': 'invalid'}))
-        self.assertFalse(form.validate())
-
-        form = TestForm(MultiDict({'date': '2012-09-06T01:29:14.107000+00:00'}))
-        self.assertTrue(form.validate())
-        self.assertEqual(
-            form.data['date'],
-            datetime.datetime(2012, 9, 6, 1, 29, 14, 107000, tzinfo=tzutc())
-        )
-
-        form = TestForm(MultiDict({'date': '2012-09-06 01:29:14'}))
-        self.assertTrue(form.validate())
-        self.assertEqual(form.data['date'], datetime.datetime(2012, 9, 6, 1, 29, 14))
 
 
 class SecretTestCase(unittest.TestCase):
